Move CliArgs to mountpoint-s3
