+++ mountpoint-s3-fs/Cargo.toml
@@ -32,7 +32,6 @@ libc = "0.2.169"
 linked-hash-map = "0.5.6"
 metrics = "0.24.1"
 nix = { version = "0.29.0", default-features = false, features = ["fs", "process", "signal", "user"] }
-owo-colors = { version = "4.1.0", features = ["supports-colors"] }
 rand = "0.8.5"
 regex = "1.11.1"
 rusqlite = { version = "0.34.0", features = ["bundled"] }
