+++ mountpoint-s3/Cargo.toml
@@ -14,10 +14,13 @@ anyhow = { version = "1.0.95", features = ["backtrace"] }
 clap = { version = "4.5.27", features = ["derive"] }
 const_format = "0.2.34"
 futures = "0.3.31"
-serde = "1.0.217"
-tracing = "0.1.41"
-serde_json = "1.0.137"
+nix = { version = "0.29.0", default-features = false, features = ["fs", "process", "signal", "user"] }
+owo-colors = { version = "4.1.0", features = ["supports-colors"] }
 regex = "1.11.1"
+serde = "1.0.217"
+serde_json = "1.0.137"
+sysinfo = "0.33.1"
+tracing = "0.1.41"
 
 [dev-dependencies]
 mountpoint-s3-client = { path = "../mountpoint-s3-client", features = ["mock"] }
@@ -28,7 +31,6 @@ aws-config = "1.5.15"
 aws-credential-types = "1.2.1"
 aws-sdk-s3 = "1.71.0"
 futures = { version = "0.3.31", features = ["thread-pool"] }
-nix = { version = "0.29.0" }
 predicates = "3.1.3"
 proptest = "1.6.0"
 rand = "0.8.5"
