+++ mountpoint-s3/src/fstab.rs
@@ -1,8 +1,10 @@
+use std::env;
+
 use anyhow::anyhow;
 use clap::error::ErrorKind;
 use clap::{Command, CommandFactory, Parser};
-use mountpoint_s3_fs::cli::CliArgs;
-use std::env;
+
+use crate::cli::CliArgs;
 
 #[derive(Parser, Debug)]
 #[clap(name = "mount-s3", disable_help_flag = true)]
